add benchmarking functionality
Benchmarking requires 2 bladeRF 2.0 micros xA9s. One bladeRF is put into receive
mode by passing the -r parameter. The other bladeRF is put transmit mode by
passing bladeRF-linux-mac80211 the -t parameter. The -t parameter specifies a
the modulation index (0 to 7 for 6Mbps through 54 Mbps). The number of packets
can be speficied with -c, by default this value is 100. The length of the packet
can be changed from 200 bytes using the -l parameter.
Both -r and -t are compatible with the -f parameter.